refactor(front): move credits DTO types out of lib/api into types/api
Eighth batch of the lib/api DTO migration: internal credits endpoints. All four
files were partial (DTOs + business logic); DTO response bodies move to
types/api/credits/*, business logic and domain error types stay.

- awu_pool_summary, members_seats: back-import their response body (function
  return type).
- programmatic_usage_limit, upgrade_requests: no back-import (functions return
  domain types/Results, not the DTOs).
- 9 consumers updated; 5 front-api routes split (DTO -> types, functions and
  domain error types -> lib).
