+++ front/lib/api/credits/members_seats.ts
@@ -9,20 +9,10 @@ import { MembershipResource } from "@app/lib/resources/membership_resource";
 import { SubscriptionResource } from "@app/lib/resources/subscription_resource";
 import { concurrentExecutor } from "@app/lib/utils/async_utils";
 import logger from "@app/logger/logger";
+import type { GetMembersSeatsResponseBody } from "@app/types/api/credits/members_seats";
 import type { MembershipSeatType } from "@app/types/memberships";
 import type { LightWorkspaceType } from "@app/types/user";
 import type { Subscription } from "@metronome/sdk/resources";
-
-export type GetMembersSeatsResponseBody = {
-  // Active members per seat type, from the DB ("assigned" seats).
-  seatTypes: Partial<Record<MembershipSeatType, number>>;
-  // Total seat quantity billed in Metronome per seat type (assigned +
-  // unassigned). Absent when the workspace isn't on a Metronome seat contract,
-  // or omitted for a seat type if Metronome couldn't be read. The UI derives
-  // the unassigned count as `metronomeSeats - seatTypes` per type.
-  metronomeSeats: Partial<Record<MembershipSeatType, number>>;
-  total: number;
-};
 
 /**
  * Read the current quantity from a subscription's `quantity_schedule` (used for
