+++ front/lib/api/credits/programmatic_usage_limit.ts
@@ -15,14 +15,6 @@ import { isCreditPricedPlan } from "@app/types/plan";
 import type { Result } from "@app/types/shared/result";
 import { Err, Ok } from "@app/types/shared/result";
 
-export type GetProgrammaticUsageLimitResponseBody = {
-  monthlyCapCredits: number | null;
-};
-
-export type PutProgrammaticUsageLimitResponseBody = {
-  monthlyCapCredits: number | null;
-};
-
 /**
  * Read the workspace's programmatic usage monthly cap.
  *
