+++ front/types/api/credits/programmatic_usage_limit.ts
@@ -1,0 +1,7 @@
+export type GetProgrammaticUsageLimitResponseBody = {
+  monthlyCapCredits: number | null;
+};
+
+export type PutProgrammaticUsageLimitResponseBody = {
+  monthlyCapCredits: number | null;
+};
